# openapi.yaml
openapi: 3.1.0
info:
  title: Auth Service API
  description: |
    API для аутентификации и авторизации пользователей.
    Сервис поддерживает регистрацию, вход, управление сессиями и токенами.
  version: 1.0.0
  contact:
    name: Diana Dvoryadkina
    email: ddvoryadkina@yandex.ru

tags:
  - name: auth
    description: Операции аутентификации и авторизации

paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: Регистрация нового пользователя
      description: |
        Регистрирует нового пользователя в системе.
        После успешной регистрации возвращает access token и устанавливает refresh token в httpOnly cookie.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: SecurePassword123!
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          headers:
            Set-Cookie:
              description: Refresh token в httpOnly cookie
              schema:
                type: string
                example: refresh_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth/refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Неверный запрос (валидация не прошла)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Validation failed"
                message: "Email is required"
        '409':
          description: Пользователь с таким email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Conflict"
                message: "User with this email already exists"
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags:
        - auth
      summary: Вход пользователя
      description: |
        Аутентифицирует пользователя по email и паролю.
        После успешного входа возвращает access token и устанавливает refresh token в httpOnly cookie.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecurePassword123!
      responses:
        '200':
          description: Успешный вход
          headers:
            Set-Cookie:
              description: Refresh token в httpOnly cookie
              schema:
                type: string
                example: refresh_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth/refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Unauthorized"
                message: "Invalid email or password"
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags:
        - auth
      summary: Обновление токенов
      description: |
        Обновляет пару токенов (access + refresh) используя текущий refresh token.
        Refresh token должен быть установлен в httpOnly cookie.
        Старый refresh token будет инвалидирован, новый будет установлен в cookie.
      operationId: refresh
      responses:
        '200':
          description: Токены успешно обновлены
          headers:
            Set-Cookie:
              description: Новый refresh token в httpOnly cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверный или истекший refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Unauthorized"
                message: "Invalid or expired refresh token"
        '400':
          description: Refresh token не найден в cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags:
        - auth
      summary: Выход пользователя
      description: |
        Выполняет выход пользователя из системы.
        Инвалидирует refresh token (добавляет в blacklist) и очищает cookie.
        Требует авторизации через access token.
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Успешный выход
          headers:
            Set-Cookie:
              description: Очистка refresh token cookie
              schema:
                type: string
                example: refresh_token=; HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth/refresh; Max-Age=0
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                message: "Logged out successfully"
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      tags:
        - auth
      summary: Получение профиля текущего пользователя
      description: |
        Возвращает информацию о текущем аутентифицированном пользователе.
        Требует валидный access token в заголовке Authorization.
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Неавторизован или неверный токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token в формате JWT.
        Токен должен быть передан в заголовке Authorization: Bearer <token>

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email пользователя
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: Пароль пользователя (минимум 8 символов)
          example: SecurePassword123!

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email пользователя
          example: user@example.com
        password:
          type: string
          format: password
          description: Пароль пользователя
          example: SecurePassword123!

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        token_type:
          type: string
          description: Тип токена
          example: Bearer
        expires_in:
          type: integer
          description: Время жизни токена в секундах
          example: 900
        user:
          $ref: '#/components/schemas/UserInfo'

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID пользователя
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          description: Email пользователя
          example: user@example.com
        created_at:
          type: string
          format: date-time
          description: Дата создания аккаунта
          example: 2024-01-15T10:30:00Z
        updated_at:
          type: string
          format: date-time
          description: Дата последнего обновления
          example: 2024-01-15T10:30:00Z
        last_login_at:
          type: string
          format: date-time
          nullable: true
          description: Дата последнего входа
          example: 2024-01-20T15:45:00Z
        is_email_verified:
          type: boolean
          description: Подтвержден ли email
          example: false

    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID пользователя
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          description: Email пользователя
          example: user@example.com

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Тип ошибки
          example: "Validation failed"
        message:
          type: string
          description: Сообщение об ошибке
          example: "Email is required"
        details:
          type: object
          description: Дополнительные детали ошибки (опционально)
          additionalProperties: true

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          description: Сообщение об успехе
          example: "Operation completed successfully"